Improve local devcontainer experience (#910)

+++ install-vfdeps.sh
@@ -1,2 +1,6 @@
-ln -sf /vf-llvm-clang-build-$VF_LLVM_CLANG_BUILD_VERSION /tmp/vf-llvm-clang-build-$VF_LLVM_CLANG_BUILD_VERSION
-ln -sf /$VFDEPS_NAME /tmp/$VFDEPS_NAME
+if [ ! -L "/tmp/vf-llvm-clang-build-$VF_LLVM_CLANG_BUILD_VERSION" ]; then
+    ln -sf /vf-llvm-clang-build-$VF_LLVM_CLANG_BUILD_VERSION /tmp/vf-llvm-clang-build-$VF_LLVM_CLANG_BUILD_VERSION
+fi
+if [ ! -L "/tmp/$VFDEPS_NAME" ]; then
+    ln -sf /$VFDEPS_NAME /tmp/$VFDEPS_NAME
+fi
